👌 Rename upgrade module to adhere to folder structure changes
