+++ upgrades/0.0.88.activate-argo-app-sync/context.go
@@ -3,7 +3,7 @@ package main
 import (
 	"context"
 
-	"github.com/oslokommune/okctl-upgrade/0.0.88.activate-argo-app-sync/pkg/logger"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.88.activate-argo-app-sync/pkg/logger"
 	"github.com/spf13/afero"
 )
