+++ upgrades/0.0.88.activate-argo-app-sync/main.go
@@ -4,12 +4,12 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"github.com/oslokommune/okctl-upgrade/0.0.88.activate-argo-app-sync/pkg/logger"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.88.activate-argo-app-sync/pkg/logger"
 	"log"
 	"os"
 	"path/filepath"
 
-	"github.com/oslokommune/okctl-upgrade/0.0.88.activate-argo-app-sync/pkg/commonerrors"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.88.activate-argo-app-sync/pkg/commonerrors"
 	"github.com/oslokommune/okctl/pkg/logging"
 	"github.com/spf13/afero"
 	"github.com/spf13/cobra"
