+++ upgrades/0.0.88.activate-argo-app-sync/pkg/kubectl/binary/api.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"io"
 
-	"github.com/oslokommune/okctl-upgrade/0.0.88.activate-argo-app-sync/pkg/kubectl"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.88.activate-argo-app-sync/pkg/kubectl"
 )
 
 // Apply runs kubectl apply on a manifest
