+++ upgrades/0.0.88.activate-argo-app-sync/upgrade.go
@@ -3,9 +3,9 @@ package main
 import (
 	"fmt"
 
-	"github.com/oslokommune/okctl-upgrade/0.0.88.activate-argo-app-sync/pkg/argocd"
-	"github.com/oslokommune/okctl-upgrade/0.0.88.activate-argo-app-sync/pkg/kubectl/binary"
-	"github.com/oslokommune/okctl-upgrade/0.0.88.activate-argo-app-sync/pkg/okctl"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.88.activate-argo-app-sync/pkg/argocd"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.88.activate-argo-app-sync/pkg/kubectl/binary"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.88.activate-argo-app-sync/pkg/okctl"
 )
 
 func upgrade(upgradeContext Context, flags cmdFlags) error {
